fix db bug

+++ logic/src/main/java/com/stasenkots/logic/network/mappers/StudentMapper.kt
@@ -2,14 +2,17 @@ package com.stasenkots.logic.network.mappers
 
 import com.stasenkots.logic.entity.student.Student
 import com.stasenkots.logic.network.dto.student.StudentResponse
+import com.stasenkots.logic.utils.toLong
+import java.text.SimpleDateFormat
+import java.time.LocalDate
 import javax.inject.Inject
 
-class StudentMapper @Inject constructor(){
-    fun map(from:StudentResponse):Student{
+class StudentMapper @Inject constructor() {
+    fun map(from: StudentResponse): Student {
         return Student(
-            id =from.objectId?:throw Exception("invalid student id"),
+            id = from.objectId ?: throw Exception("invalid student id"),
             name = from.name.orEmpty(),
-            updatedAt = from.updatedAt.orEmpty()
+            updatedAt = from.updatedAt.orEmpty().toLong()
         )
     }
 }
